Add pointer support
Only absolute mode for now.  Our focus behavior is a bit novel too.  Instead
of passing all pointer events to the guest, we avoid relaying pointer events
to the guest unless the drawing area widget has focus.

As long as the widget does not have focus, we do not hide the host cursor.