Hook the unrealize event for the display widget in core.c to suppress Display Change Listener notifications after the widget is no longer writable.
Unrealize happens when the gdk resources associated with a widget is freed.
Since the DCL notifications result in changes happening to the gdk resources,
this is a pretty good place to hook into to unregister the DCL.

NB, the whole valid thing is a hack.  You cannot currently unregister a DCL
which is silly.