virtio: Add memory statistics reporting to the balloon driver
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1268947268-21943-4-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 7918
O-Subject: [PATCH 3/3] virtio: Add memory statistics reporting to the balloon
	driver
Bugzilla: 574939
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

This is the backport of the original upstream commit:

625a5befc2e3200b396594f002218d235e375da5

Plus the following fixes:

5c366a8a3d7ac71beda8499caa815cb3ea95eb58
93d67ee69d6104976961fbc98654872de28a5c96
6d0ee85040e4d238e2483191fe6e74aebbecd5d5

As this is an important change with QMP effects, its original
commit log follows:

When using ballooning to manage overcommitted memory on a host, a system for
guests to communicate their memory usage to the host can provide information
that will minimize the impact of ballooning on the guests.  The current method
employs a daemon running in each guest that communicates memory statistics to a
host daemon at a specified time interval.  The host daemon aggregates this
information and inflates and/or deflates balloons according to the level of
host memory pressure.  This approach is effective but overly complex since a
daemon must be installed inside each guest and coordinated to communicate with
the host.  A simpler approach is to collect memory statistics in the virtio
balloon driver and communicate them directly to the hypervisor.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
---
 balloon.h           |    7 ++-
 hw/virtio-balloon.c |  120 ++++++++++++++++++++++++++++++++++++++++++++++++--
 hw/virtio-balloon.h |   15 ++++++
 monitor.c           |   99 ++++++++++++++++++++++++++----------------
 qemu-monitor.hx     |    3 +-
 vl.c                |   21 ++++++---
 6 files changed, 212 insertions(+), 53 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>