Replace calls of old bdrv_open
RH-Author: Kevin Wolf <kwolf@redhat.com>
Message-id: <1270640579-13548-6-git-send-email-kwolf@redhat.com>
Patchwork-id: 8418
O-Subject: [RHEL-6 KVM PATCH 5/5] Replace calls of old bdrv_open
Bugzilla: 579974
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

Bugzilla: 579974
Upstream status: Submitted

What is known today as bdrv_open2 becomes the new bdrv_open. All remaining
callers of the old function are converted to the new one. In some places they
even know the right format, so they should have used bdrv_open2 from the
beginning.

One of these places is the metadata preallocation in qcow2. If you created a
qcow2 image on a host device, it would open it as a host_device instead of
qcow2 and consequently fail when qcow2 tries to use the BlockDriverState as its
own.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
---
 block.c       |   22 ++++++++--------------
 block.h       |    5 ++---
 block/qcow2.c |    4 ++--
 block/vmdk.c  |    2 +-
 block/vvfat.c |    5 ++++-
 hw/xen_disk.c |    2 +-
 monitor.c     |    2 +-
 qemu-img.c    |   16 ++++++++--------
 qemu-io.c     |    2 +-
 qemu-nbd.c    |    2 +-
 vl.c          |    2 +-
 11 files changed, 30 insertions(+), 34 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>