virtio-serial: save/load: Ensure nr_ports on src and dest are same.
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <1272359264-8464-3-git-send-email-amit.shah@redhat.com>
Patchwork-id: 8848
O-Subject: [RHEL6 PATCH v4 02/17] virtio-serial: save/load: Ensure nr_ports on
	src and dest are same.
Bugzilla: 574296
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Alon Levy <alevy@redhat.com>

The number of ports on the source as well as the destination machines
should match. If they don't, it means some ports that got hotplugged on
the source aren't instantiated on the destination. Or that ports that
were hot-unplugged on the source are created on the destination.

Bugzilla: 574296
Upstream: <posted>

Signed-off-by: Amit Shah <amit.shah@redhat.com>
Reported-by: Juan Quintela <quintela@redhat.com>
---
 hw/virtio-serial-bus.c |   18 ++++++++++++++++--
 1 files changed, 16 insertions(+), 2 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>