virtio-serial: save/load: Send target host connection status if different
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <1272359264-8464-5-git-send-email-amit.shah@redhat.com>
Patchwork-id: 8850
O-Subject: [RHEL6 PATCH v4 04/17] virtio-serial: save/load: Send target host
	connection status if different
Bugzilla: 574296
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Alon Levy <alevy@redhat.com>

If the host connection to a port is closed on the destination machine
after migration, whereas the connection was open on the source, the
guest has to be informed of that.

Similar for a host connection open on the destination.

Bugzilla: 574296
Upstream: <posted>

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/virtio-serial-bus.c |   11 +++++++++++
 1 files changed, 11 insertions(+), 0 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>