virtio-serial: Use control messages to notify guest of new ports
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <1272359264-8464-6-git-send-email-amit.shah@redhat.com>
Patchwork-id: 8851
O-Subject: [RHEL6 PATCH v4 05/17] virtio-serial: Use control messages to notify
	guest of new ports
Bugzilla: 574296
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Alon Levy <alevy@redhat.com>

Allow the port 'id's to be set by a user on the command line. This is
needed by management apps that will want a stable port numbering scheme
for hot-plug/unplug and migration.

Since the port numbers are shared with the guest (to identify ports in
control messages), we just send a control message to the guest
indicating addition of new ports (hot-plug) or notifying the guest of
the available ports when the guest sends us a DEVICE_READY control
message.

Bugzilla: 574296
Upstream: <posted>

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/virtio-console.c    |    2 +
 hw/virtio-serial-bus.c |  184 +++++++++++++++++++++++++++++++----------------
 hw/virtio-serial.h     |   17 +++--
 3 files changed, 133 insertions(+), 70 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>