virtio-serial: Handle output from guest to unintialised ports
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <1272446224-773-1-git-send-email-amit.shah@redhat.com>
Patchwork-id: 8888
O-Subject: [RHEL6 PATCH v4.1 18/19] virtio-serial: Handle output from guest to
	unintialised ports
Bugzilla: 574296
RH-Acked-by: Alon Levy <alevy@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

Guests can send data to ports that the host hasn't initialised. This
used to be handled properly earlier but the code changed in the last
series that made this condition expose a segfault. Fix that.

(This patch depends on the previous series)

Upstream: <posted>
Bugzilla: 574296

Reported-by: Alon Levy <alevy@redhat.com>
Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/virtio-serial-bus.c |   18 +++++++++++++-----
 1 files changed, 13 insertions(+), 5 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>