blkdebug: Add events and rules
RH-Author: Kevin Wolf <kwolf@redhat.com>
Message-id: <1272978696-18996-7-git-send-email-kwolf@redhat.com>
Patchwork-id: 8992
O-Subject: [RHEL-6 qemu-kvm PATCH 6/7] blkdebug: Add events and rules
Bugzilla: 588756
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

Bugzilla: 588756
Upstream commit: 8b9b0cc2fd1b866c0ce6c7f7385d840aad8b4c2c

Block drivers can trigger a blkdebug event whenever they reach a place where it
could be useful to inject an error for testing/debugging purposes.

Rules are read from a blkdebug config file and describe which action is taken
when an event is triggered. For now this is only injecting an error (with a few
options) or changing the state (which is an integer). Rules can be declared to
be active only in a specific state; this way later rules can distiguish on
which path we came to trigger their event.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
---
 block.c          |   12 +++
 block.h          |    9 ++
 block/blkdebug.c |  250 +++++++++++++++++++++++++++++++++++++++++++++++++++++-
 block_int.h      |    2 +
 4 files changed, 272 insertions(+), 1 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>