block: Add wr_highest_sector blockstat
RH-Author: Kevin Wolf <kwolf@redhat.com>
Message-id: <1273512519-16871-4-git-send-email-kwolf@redhat.com>
Patchwork-id: 9138
O-Subject: [RHEL-6 qemu-kvm PATCH 3/3] block: Add wr_highest_sector blockstat
Bugzilla: 590998
RH-Acked-by: Luiz Capitulino <lcapitulino@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Christoph Hellwig <chellwig@redhat.com>

This adds the wr_highest_sector blockstat which implements what is generally
known as the high watermark. It is the highest offset of a sector written to
the respective BlockDriverState since it has been opened.

The query-blockstat QMP command is extended to add this value to the result,
and also to add the statistics of the underlying protocol in a new "parent"
field. Note that to get the "high watermark" of a qcow2 image, you need to look
into the wr_highest_sector field of the parent (which can be a file, a
host_device, ...). The wr_highest_sector of the qcow2 BlockDriverState itself
is the highest offset on the _virtual_ disk that the guest has written to.

Signed-off-by: Kevin Wolf <kwolf@redhat.com>
(cherry picked from commit 294cc35f3db84017e0c176c15dd9a1b711e8354d)
---
 block.c     |   73 ++++++++++++++++++++++++++++++++++++++++++++++------------
 block_int.h |    1 +
 2 files changed, 59 insertions(+), 15 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>