QMP: Add error reason to BLOCK_IO_ERROR event
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1273877360-19825-3-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 9296
O-Subject: [PATCH 2/2] QMP: Add error reason to BLOCK_IO_ERROR event
Bugzilla: 586349
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>

Bugzilla: 586349
Upstream: N/A - Explained below

This commit adds a new member to the BLOCK_IO_ERROR event, which
exposes the reason of the error as a named errno. That's, it
can be: "eperm", "eio", "enospc" or "eother".

This is a RHEL6-only private extension, just to satisfy vdsm needs.

Upstream has rejected using named errno in QMP and we're still
discussing what's the best solution for this.

However, Anthony has ACKed using a private extension for this, as
RHEL6 (probably) can't afford to wait for the upstream solution.

NOTE: Only tested the basic ENOSPC case with an ide drive.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
---
 QMP/qmp-events.txt |    9 ++++++++-
 block.c            |   28 +++++++++++++++++++++++++++-
 block.h            |    2 +-
 hw/ide/core.c      |    6 +++---
 hw/scsi-disk.c     |    6 +++---
 hw/virtio-blk.c    |    6 +++---
 6 files changed, 45 insertions(+), 12 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>