VM can't read files to start
RH-Author: Eduardo Habkost <ehabkost@redhat.com>
Message-id: <20100524160607.GD20497@blackpad.lan.raisama.net>
Patchwork-id: 9500
O-Subject: Re: VM can't read files to start
Bugzilla: 593909
RH-Acked-by: Juan Quintela <quintela@redhat.com>

On Mon, May 24, 2010 at 01:19:39PM +0200, Juan Quintela wrote:
>
> bugzilla:
> Bug 593909 - VM can not start by using qemu-kvm-0.12.1.2-2.56.el6
>
> We need to blacklist "file" block format, as we changed the "raw" name
> to "file" everywhere else.

Are you sure it is safe to remove "raw" from the whitelist?

I will change it to append "file" to the list instead of removing "raw"
from the whitelist.

This is the patch I plan to include. Please ACK.

---

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>