device-assignment: Don't use libpci
RH-Author: Alex Williamson <alex.williamson@redhat.com>
Message-id: <20100524200102.23875.49888.stgit@virtlab9.virt.bos.redhat.com>
Patchwork-id: 9515
O-Subject: [RHEL6.0 qemu-kvm PATCH 2/3] device-assignment: Don't use libpci
Bugzilla: 595495
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>

Bugzilla: 595495
Upstream status: submitted

We've already got an open fd for PCI config space for the device, we
might as well use it.  This also makes sure that if we're making use of
a privileged file descriptor opened for us, we use it for all accesses
to the device.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Chris Wright <chrisw@redhat.com>
---

 configure              |   14 +++++-----
 hw/device-assignment.c |   68 ++++++++++++++++++++++++++++++++++++------------
 hw/device-assignment.h |    1 -
 3 files changed, 58 insertions(+), 25 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>