vhost-net: fix reversed logic in mask notifiers
RH-Author: Michael S. Tsirkin <mst@redhat.com>
Message-id: <20100525141420.GA26016@redhat.com>
Patchwork-id: 9538
O-Subject: [RHEL6.0 PATCH] vhost-net: fix reversed logic in mask notifiers
Bugzilla: 585940
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Amit Shah <amit.shah@redhat.com>

When guest notifier is assigned, we set mask notifier, which will assign
kvm irqfd.  When guest notifier is unassigned, mask notifier is unset,
which should unassign kvm irqfd.

The way to do this is to call mask notifier telling it to mask the vector.
This, unless vector is already masked which unassigns irqfd already.

The logic in unassign was reversed, which left kvm irqfd assigned.

This patch is qemu-kvm only as irqfd is not upstream.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Reported-by: Amit Shah <amit.shah@redhat.com>

Bugzilla: 585940
(this completes the fix)
Upstream status: posted
Brew build: https://brewweb.devel.redhat.com/taskinfo?taskID=2469235

---
 hw/msix.c |    4 +++-
 1 files changed, 3 insertions(+), 1 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>