Fix console_write_ch on 64-bit big-endian hosts
Currently, console_ch_t is defined as an unsigned long.  However, immediately
after it's definition, we treat it as a uint32_t *.  This will work on a little
endian system because of the way bits are layed out but will fail miserably
on big endian hosts.

This patch fixes the code to do the correct thing.  This addresses
https://bugs.launchpad.net/qemu/+bug/568614

Reported-by: Devin J. Pohly
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>