virtio: Use ioeventfd for virtqueue notify
Virtqueue notify is performed synchronously with the VM unable to
execute code while hardware emulation runs.  On systems that support
KVM, the ioeventfd mechanism can be used to make virtqueue notify a
lightweight exit by deferring hardware emulation to the iothread and
allowing the VM to continue execution.

This patch takes advantage of ioeventfd virtqueue notify for all virtio
devices.  Linux kernels pre-2.6.34 only allow for 6 ioeventfds per VM so
care must be taken so that vhost-net, the other ioeventfd user in QEMU,
is able to function.  On such kernels, do not use ioeventfd virtqueue
notify.

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>