qdev: split out device_add code and monitor code
This is a big patch because things are pretty tangelled up right now.

The theory is that qdev should be able to build without any real dependencies.
Certainly, without the Monitor as a dependency.  There are a few problems that
prevent this.  The first is that print_dev is part of BusInfo which means
that ever bus is supposed to implement Monitor code.

This is a bad design because it means that the Monitor is not in control of its
output because it's spread across a wide variety of devices.  In general, we
should strive to centralize the Monitor code and ensure that it interacts with
the device models in a controlled way.

Eliminating the print_dev is pretty straight forward but it introduces a build
dependency on target_addr_t.  Instead of making qdev build with target_addr_t,
I decided to just go ahead and split the code at once.

A lot of the non qdev core code was also device_add support which is really
a UI element.  Again, we don't want this stuff in core qdev because we want
it to all work through well defined interfaces.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>