qdev: refactor hotplug to not be such a special case
This patch attempts to make hotplug more natural as far as qdev is concerned.

The first thing we do is eliminate the global notion of "machine done" that's
used to enforce hotplug today.  Instead, we add another device state called
REALIZED which indicates that the device is now visible to a guest.  The
REALIZED state is generally useful as an indication that we know have to be
careful about changes we make to the device and to the tree.

Instead of having device's react different if they're "hotplugged", we introduce
a bus-level callback that allows a bus to accept or reject a device when it's
added or removed.  Hotplug is just a special case which is when a device is
added after the bus device has already been realized.  This interface is
generally useful because it gives busses a graceful way to reject devices due
to finite resources of the bus or because of other bus restrictions.  We don't
really exploit this yet but that's the topic of a future patch.

There's a few ugly places that need to be hot plug aware.  We handle that by
checking if the parent bus is realized.  We need to refactor these away.

One small change this series introduces is that hotplug based on pci_add no
longer throws the same error message if a bus doesn't support hotplug.  There's
really no way to do this gracefully given how pci_add works.  This command
should be deprecated though so I don't think this is a problem in practice.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>