qdev: switch use of qdev_unplug to qdev_free
There's no need to have an explicit unplug operation.  Instead, we just use
qdev_free and have it do the right thing depending on what the device's state
is.

To make this work, we have to let qdev_free fail.  This is unavoidable though
as the practical semantics of doing a qdev_free on a live device tree is
failure anyway without an unplug.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>