qdev: remove qdev_create(NULL, ...) hack
Introduce a singleton interface for the main system bus and convert all users
of qdev_create(NULL, ...) to use this interface.  We really want to eliminate
this singleton interface in the long run but this is far better than having
the core of qdev create an implicit bus.

Sysbus is a bit broken today.  It takes target_phys_addr_t's in it's interface
which means that you cannot build anything that includes sysbus.h in libhw64.
We work around this using another header but the proper solution is to refactor
sysbus to use a bus-specific type.  That's a future patch though.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>